REFACTOR: move class ModelInspector to `.inspection.base` sub-package
