+++ src/facet/inspection/__init__.py
@@ -8,4 +8,3 @@ of a learner pipeline which has been fitted using cross-validation.
 from ._function_inspector import *
 from ._inspection import *
 from ._learner_inspector import *
-from ._model_inspector import *
