+++ src/facet/inspection/base/__init__.py
@@ -1,0 +1,5 @@
+"""
+Abstract base classes for model inspection.
+"""
+
+from ._model_inspector import *
